lumi-v2-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
loguru==0.7.2
supabase==2.11.0
email-validator==2.2.0
qdrant-client==1.12.1
llama-index==0.12.5
llama-index-vector-stores-qdrant==0.4.0
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
sentence-transformers==3.4.1
groq==0.11.0
llama-index-embeddings-huggingface==0.5.1
httpx==0.27.2